requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
yt-dlp
moviepy
pydantic